Extension: move auth files to follow components structure
