+++ front/extension/app/main.tsx
@@ -5,11 +5,10 @@ import "@dust-tt/sparkle/dist/sparkle.css";
 // Local tailwind components override sparkle styles
 import "@app/styles/components.css";
 
+import { AuthProvider } from "@app/extension/app/src/components/auth/AuthProvider";
+import { routes } from "@app/extension/app/src/routes";
 import ReactDOM from "react-dom/client";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
-
-import { AuthProvider } from "./src/context/AuthProvider";
-import { routes } from "./src/routes";
 
 const router = createBrowserRouter(routes);
